feat: added connected page
